+++ commands.js
@@ -6,6 +6,8 @@ const taskDeleteHandler = require("./commands/taskDelete");
 const tasksHandler = require("./commands/tasks");
 
 const reminderActionsHandler = require("./commands/reminderActions");
+
+const addAdminHandler = require("./commands/addadmin");
 const listAdminsHandler = require("./commands/listadmins");
 
 module.exports = function (app, db) {
@@ -15,5 +17,6 @@ module.exports = function (app, db) {
   taskDeleteHandler(app, db);
   tasksHandler(app, db);
   reminderActionsHandler(app, db);
+  addAdminHandler(app, db);
   listAdminsHandler(app, db);
 };